IOS7: Refine the touch controls even further
Correct the speed of the mouse pointer to be the same as the finger
speed.

To be able to move the pointer without triggering a click, add a touch
timer that can be invalidated if the one finger touch is moved. If the
one finger touch is moved before the timer is triggered the left mouse
button event will be canceled.