CREATE_PROJECT: Use xcframeworks for ios7 libraries
The static libraries, required for building the ios7 port, have
been ported to use the Apple xcframework which can contain static
libraries for different targets. Using xcframeworks enables the
possibility to have static libraries for arm64 for both iPhoneOS
and iPhoneSimulator.

This commit require the user to change their libraries to the new
scummvm-ios7-libs-v3.