SCUMM: Fix MIDI music for Fate of Atlantis PPC version
The midi files for Indiana Jones and the fate of Atlantis are in the
PPC version packed in the file "Fate of Atlantis Data". This file is
set to be in version 1 of the Macintosh iMuse driver. This is defined
in:
	static const char *const fileNames[3][4] = {
		{
			"iMUSE Setups",
			nullptr,
			nullptr,
			nullptr
		},
		{
			"Instruments",
			"Day of the Tentacle Demo",
			"Day of the Tentacle",
			"Fate of Atlantis PowerPC"
		},
		{
			"Instruments",
			"Sam & Max Demo",
			"Sam & Max",
			nullptr
		}

Change the version to 1 for the INDY4 PPC version to get MIDI music
working.